+++ siwfood/src/main/java/it/uniroma3/siwfood/controller/RicettaController.java
@@ -19,6 +19,7 @@ import it.uniroma3.siwfood.model.Cuoco;
 import it.uniroma3.siwfood.model.Immagine;
 import it.uniroma3.siwfood.model.Ingrediente;
 import it.uniroma3.siwfood.model.Ricetta;
+import it.uniroma3.siwfood.model.User;
 import it.uniroma3.siwfood.service.CuocoService;
 import it.uniroma3.siwfood.service.ImmagineService;
 import it.uniroma3.siwfood.service.IngredienteService;
@@ -31,7 +32,7 @@ import jakarta.validation.Valid;
 
 
 @Controller
-public class RicettaController {
+public class RicettaController extends GlobalController{
 	
 	@Autowired RicettaService ricettaService;
 	@Autowired IngredienteService ingredienteService;
@@ -57,6 +58,12 @@ public class RicettaController {
 	//gestisce la visualizzazione del form
 	@GetMapping("/chef/formNewRicetta/{cuoco_id}")
 	public String formNewRicetta(@PathVariable("cuoco_id") Long id,Model model) {
+		User user = getCredentials().getUser();
+        if ((!getCredentials().isAdmin()
+                && cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != id)
+                || cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != id) {
+            return "redirect:/error";
+        }
 		model.addAttribute("cuoco", this.cuocoService.findById(id));
 		model.addAttribute("ricetta", new Ricetta());
 		return "formNewRicetta.html";
@@ -85,17 +92,38 @@ public class RicettaController {
 	
 	//gestisce la visualizzazione del form
 	@GetMapping("/chef/formNewIngrediente/{ricetta_id}")
-	public String formNewIngrediente(@PathVariable("ricetta_id") Long id,Model model) {
-		model.addAttribute("ricetta", this.ricettaService.findById(id));
+	public String formNewIngrediente(@PathVariable("ricetta_id") Long ricettaId, Model model) {
+		Ricetta ricetta = this.ricettaService.findById(ricettaId);
+		Cuoco cuoco = ricetta.getCuoco();
+		Long cuocoId = cuoco.getId();
+		
+		User user = getCredentials().getUser();
+		if ((!getCredentials().isAdmin()
+				&& cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != cuocoId)
+				|| cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != cuocoId) {
+			return "redirect:/error";
+		}
+		
+		model.addAttribute("ricetta", ricetta);
 		model.addAttribute("ingrediente", new Ingrediente());
 		return "formNewIngrediente.html";
 	}
 
 	//gestisce il salvataggio della ricetta e il reindirizzamento alla pagina della ricetta appena creata
 	@PostMapping("/chef/saveIngredienti/{ricetta_id}")
-	public String newIngredienti(/*@Valid*/ @PathVariable("ricetta_id") Long id, @ModelAttribute Ingrediente ingrediente/* ,BindingResult bindingResult*/) {
+	public String newIngredienti(/*@Valid*/ @PathVariable("ricetta_id") Long id, @ModelAttribute Ingrediente ingrediente,@RequestParam("immagine") MultipartFile immagine) throws IOException {/* ,BindingResult bindingResult*/
 		//this.ingredienteValidator.validate(ingrediente, bindingResult);
 		//if (!bindingResult.hasErrors()) {
+			if (!immagine.isEmpty()) {
+				Immagine img = new Immagine();
+				img.setFileName(immagine.getOriginalFilename());
+				img.setImageData(immagine.getBytes());
+				if (ingrediente.getImmagini() == null) {
+					ingrediente.setImmagini(new ArrayList<>());
+				}
+				ingrediente.getImmagini().add(img);
+				immagineService.save(img);
+			}
 			Ricetta ricetta = this.ricettaService.findById(id);
 			ingrediente.setRicetta(ricetta);
 			this.ingredienteService.save(ingrediente);
@@ -126,12 +154,27 @@ public class RicettaController {
 	//@PathVariable Estrae variabili di percorso dall'URL della richiesta.
 	@PostMapping("/chef/deletericetta/{id}")
     public String deleteRicetta(@PathVariable Long id) {
+		User user = getCredentials().getUser();
+		if ((!getCredentials().isAdmin()
+                && cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != id)
+                || cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != id) {
+            return "redirect:/error";
+        }
         ricettaService.deleteRicettaById(id);
         return "redirect:/ricette"; // Redirect alla lista delle ricette dopo la cancellazione
     }
 	
 	@GetMapping("/chef/editricetta/{id}")  //Mappa le richieste GET all'URL /ricetta/update/{id}.
     public String getUpdateForm(@PathVariable("id") Long id, Model model) {
+		Ricetta ricetta = this.ricettaService.findById(id);
+        User user = getCredentials().getUser();
+        if ((!getCredentials().isAdmin()
+                && cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != ricetta.getCuoco()
+                        .getId())
+                || cuocoService.findbyNomeCognome(user.getName(), user.getSurname()).getId() != ricetta.getCuoco()
+                        .getId()) {
+            return "redirect:/error";
+        }
     	model.addAttribute("ricetta", this.ricettaService.findById(id)); 
         return "formUpdateRicetta.html"; //Indica il nome del template che Spring deve utilizzare per generare la risposta HTML.
 	}
